requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
pydantic==2.12.5
sqlalchemy==2.0.41
alembic==1.16.1
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-jose[cryptography]==3.3.0
python-dotenv==1.1.0
pydantic-settings==2.9.1
